Standardize API return format to array with 'code' and 'data' fields
An implementation detail worth noting is that I got rid of the notion
of last_status_code. It's only present use was for setting the 'code'
field in certain API actions. Saving that information to an instance
variable was unecessary and when it's possible to avoid having state
in a service, I think it is good to do so.